Add Apple Mapkit limited provider support
Instead of adding the full token handling, this allows to directly set
an access key (usually valid for 30 min), to use this provider for
specific tiles.

To get this key, there is no other way than having a paid developer
account and follow the full procedure to generate a key. But if you can
use it, it can provide better quality tiles in some regions of the world